refactor: move News component to own folder
